Implement basic functionalities
Now you can move the camera, draw tiles and remove tiles

+++ core/src/com/actionteam/geometryadventures/mapcreator/model/Tile.java
@@ -10,5 +10,5 @@ public class Tile {
     public int textureIndex;
     public float x;
     public float y;
-    public float z;
+    public int z;
 }
